Add ESLint rule to require function return types to be specified if exported
I only want to enforce this in utility as it feels unnecessary in packages such as components or AlexCLine where functions aren't necessarily the main purpose (components almost always returns a ReactNode and AlexCLine isn't even a tool to be used in TypeScript/JavaScript code). But with utility, the aim is to provide functions to be used, and as such, we want to be as explicit as possible about how to use them. Explicitly specifying the return type helps with this.