Add test for if the day is the same, but Date instance is different
The existing true test only checked that it returns true if we give the same date instance twice. This change adds a test that points to the same day of today, but different date instances, so that we can verify that two dates that are the same still return true, even if they're different instances of Date.