Add interpolateObjects function
This can be used in cases where we want to interpolate an object in a template string. Usually this would give [object Object] if you try and do this the usual way, but with this tagged template function, it stringifies the object for you and actually prints its contents.

This is also a nice way to practice writing tagged template string functions, which is cool.