Delete fillArrayAsync test since that function is now deprecated
Since we are no longer officially supporting fillArrayAsync, we probably don't need to test it either. It always took the longest to complete because of the extra wait() functions used in testing to justify the use of async functions, which, even though it was only 300ms, it's still longer than other utilities.

The tests for fillArray speed this up so that we only have to test the async variant once as opposed to three different times, further justifying the fillArrayAsync deprecation. Plus, if someone truly were to use fillArrayAsync from this point on, they won't even have the confidence of tests backing them up anymore, so it'd be on them at this point.