Add NonUndefined type
This is useful for checking that a type is specifically not undefined. NonNullable does exist but also checks null. Null can often be a valid response in some APIs, whereas undefined can be treated differently, so it's nice to be able to have a type that only checks for non-undefined.