Add stringToBoolean function
This helps makes boolean comparisons feel a little more airtight. We usually do this by checking if the variable is strictly the string true, which doesn't take into account the fact that it could be coming in as some other value beyond true/false. This function only returns a valid boolean if a valid boolean string is provided, otherwise it throws an error.