Add parseIntStrict function
This is useful in cases where we may want to throw an error if a given string does not represent an integer. The built-in parseInt function only returns NaN, which has the potential to be a bit screwy, especially since NaN is actually typed as a number. A hard fail parse function ensures that logic cannot continue unless the input can definitely be parsed to the expected type.