Add removeDuplicates function
This one's a classic - I remember it being one of the first challenges in bootcamp in which I did kinda struggle with. My solution ended up being to loop three times at the time, but now I only need to loop once!

Anyway, this actually has a use case beyond just a bootcamp problem now - the error mapping in Neurosongs 2 takes in some default API errors along with some custom ones, but as of now, it only loops through the custom ones. As such, I need to create a new array of the shared error codes without any duplication, which very much sounds like a job for removeDuplicates.