Rename HTTPErrorCodes to HTTPErrorCode
The type name is really only meant to represent a single one of these error code. It reads nicer in variable declarations, since you'd read something like const errorCode: HTTPErrorCode as 'the constant errorCode is an HTTP Error Code', which feels more natural in singular than it does in plural.