Add stringListToArray function
This one comes up a fair amount, especially when trying to parse .env files where we can't really set an environment variable to an array, so we have to provide a comma-separated list and split it manually. This function takes care of this automatically for us, also allowing us to configure things like the separator and whether or not to trim out whitespace.