Use Parameters type to type the parseIntStrict arguments
This ensures that the argument types for the function exactly match what parseInt expects. This Parameters type in general seems like it'd be especially useful for wrapper functions, allowing us to wrap around functions without constantly having to manually figure out and specify the parameter types of every single argument.