Use parseIntStrict to check if numbers are integers
This gets rid of the need to keep checking numbers modulo 1 everywhere, and allows parseIntStrict to automatically throw an error if an integer is not provided.

Note that this means that the error code has changed from NON_INTEGER_INPUT(S) to INTEGER_PARSING_ERROR for functions that previously checked modulo 1, and consuming logic will have to update their code to account for this accordingly.