Add tests for the utility types
Yes, we can test compile-time type logic. It's so cursed, but also so amazing at the same time. It takes some getting used to because type information only exists at compile-time, so the tests will always pass if you run npm test - it's just that the TypeScript compiler does the type assertions and will error if the types aren't what it expects them to be.

Since these tests were included retroactively, it's not the most ideal workflow in this case. Usually when writing tests, we want to watch them fail first, then make them pass. This feels especially important for type tests because they always pass in runtime, which increases the chance of false positives. However, I have tried to make them fail first when writing the tests to ensure that it works as intended.