Add createFormData function to map objects to formData
This helps make form data creation a lot less verbose and more readable, replacing a bunch of formData.append() invocations with a nicer-to-read object.

Note that, even though it should theoretically be possible to upload a Blob so we can append file contents, it does not work when we try testing it in Vitest for some reason. I think this is because of weird inconsistent behaviour between Vitest and the actual browser environment, so unfortunately, the best we can do for now is just to assert the structure and not the value.