Add a generic type to createFormData
It does absolutely nothing in the actual function types themselves, but it does help to provide type hints/autocomplete when writing out the function argument, so... there's that.

FormData is so weird because it literally sucks to use in every single way except for its ability to natively handle files and blobs and pass them over to back-end servers. Other than that, it's a very stupid feature to use and is inferior to JSON in every other way.