Deal with stringification of values natively, including support for undefined and null
This is helpful so that consuming code doesn't have to manually remember to stringify all values - the function can do it for us.

Note that this requires special treatment of undefined and null, since before they'd just be passed directly passed into either JSON.stringify or the String() constructor. Although this may be the desired behaviour sometimes, it might not always be. As such, I've provided three options - empty, stringify, and omit (defaulting to empty). Empty maps these to an empty string, stringify stringifies them, and omit leaves them out.

And yes, the roast towards pure JavaScript users is absolutely necessary.