Allow for arrays to either be resolved as multiple items, or stringified
The default is to stringify as that's the safest thing to do if you have arrays or objects. However, for things like files, it's often better to send it through as multiple repeated properties. This change allows you to choose the method of resolving arrays for your given use case.