Allow arrays of primitives and files to be added as multiple keys within FormData
FormData does allow you to add multiple items that share the same key, for some reason, so we should probably take that into account with our utility to help us create form data. Files in particular are better uploaded this way and it may help for some primitives. Note, however, that objects/arrays are not supported this way due to the amount of nesting involved.