Do not allow arrays of Blobs to be passed at the same time as arrayResolution: stringify
The whole selling point of FormData is that they can deal with Blobs directly, so best to just always pass it in directly wherever possible. Note thta this means for now, passing in no arrayResolution at the same time as a blob array will always fail as the default is always to stringify. This will be fixed in a future version, but for now, I think it's more important to prevent the accidental stringification of blob arrays.