Add getRecordKeys function.
Object.keys() exists, but it allows for anything because everything in JavaScript is an object. This ONLY works with Records (that is, objects where we define the keys/values explicitly) and ensures that the return type does correspond to the actual keys of that record.

This is often useful to use with the omitProperties function, where the second argument is an array typed as the array of keys you want to omit. As such, just passing Object.keys() here would not work as that's just typed as string[].