Add paralleliseArrays function
This works similarly to how zip works in Python, in the sense that it creates a new array with the two input arrays parallelised such that each item of the new array is an array of items at that index of the input arrays. It loops for as long as the first array if the lengths do not match, whereby leftover second array properties just become undefined.

This is useful for looping through arrays in parallel so that we no longer need to keep looping over the index of an arbitrarily chosen array. Instead we can just use this function and pass in the two arrays we want to loop through, putting the one we want to loop through all properties of at the start. This is something that is very commonly done in tagged template string functions, where we want to loop through the strings and interpolations in parallel.