Add stripIndents function
This more or less mirrors what stripIndents from common-tags does, except now it's in-house so I can control the behaviour of it more. My implementation also preserves the tab spacing, though, whereas I'm pretty sure last time I tried stripIndents from common-tags, it didn't always tend to.

Also note that the assumed default of 12 is what it seems to be set to in my editor. It may differ between editors, but I'm not too sure. If it does, I will change this in a later implementation to allow the default whitespace to be configured.