Allow stripIndents to take options
If we give the function an overload, we can enforce at the type level that we can either take StripIndentsOptions as an argument, or a string and interpolations (equivalent to passing in a template string directly). This allows us to then specify things like the whitespace length or whether or not to preserve tabs before we then pass in the template string. Essentially, it can either take a template string that returns a string, or an options argument that returns a function that takes a template string that returns a string.