Programatically calculate the whitespace length based on the first non-blank line
The way that extra whitespace works in multi-line strings is that extra indentation is added based on the next line's position in the editor. That is, if a line starts at column 4, the extra whitespace will be 4 characters long. Likewise, if a line starts at column 12, there will be 12 extra whitespace characters. This change accounts for this rather than rely on the user to set the whitespaceLength themselves.