Add a create-local-package script
This allows other packages to run against a local version of this package. While not as useful here as it is in other packages given that utility functions are often the most rigourously tested, it is helpful in cases where we may be changing the build process and want to verify it works locally first before publishing.