Use regular Readonly for deepFreeze again
It was breaking ESLint plugin as I think it was also trying to say that primitive values in objects are also readonly, for whatever reason. As such, for now I will just go back to using the regular Readonly type until I figure out what is breaking DeepReadonly in ESLint plugin.