Add JSDoc comments to a few functions
I think we should start making it a requirement in all my packages that any exported functions must have JSDoc comments, as that will help improve the developer experience by giving developers immediate guidance on how to use these utilities. That said, I have admittedly been putting this off a bit, so I am currently going through them all to add these in.

Note that for overloaded functions, this is a bit suboptimal as I have to provide JSDoc comments on each individual overload, which is a bit annoying. Not sure if there's a better way, though.