Add mutation tests to the date utilities
The Date object can be mutated directly, for whatever reason, and it is easy to get caught out. As such, I have added tests to ensure that any input dates do not get mutated.

Note that Object.freeze() does not quite work as intended, so we must assert that the value is unchanged and that the reference in memory is the same.