Deprecate the types that only exist to infer basic Zod schemas
It's not that difficult to just get those types from Zod itself, so I don't think there's much value in abstracting just that in my own utility package. What might be more useful would be if I were to have a pure Email type that existed only at compile-time, for example. Likewise with UUID, but I think a UUID type is already common in a lot of packages.