Add JSDoc comments to the tagged template utilities
This should provide a better developer experience for people using these shared components, as it gives usage information right there in the editor.

It is slightly weird, though, because I'm not sure we get much out of constantly annotating strings and interpolations when the most likely use case is to pass in a template string directly, but I tried as best as I could to annotate these.