Update @alextheman/eslint-plugin
Note that eslint-plugin-perfectionist had to be installed to this repository manually, as otherwise I think PNPM tries to resolve it as if it is at v4 as that is still the lower bound so as not to break other usage. Eventually I will tighten this up, but for now, this must be done.