Add the initial implementation of the VersionNumber class
This should make the versioning utilities feel more unified, reducing the need to keep calling parseVersion() everywhere. Now it is only ever parsed once with the constructor, and we can call methods on the unified VersionNumber to immediately operate on the known parsed version. Also, since it is a type, it can be used as a type annotation for better typing should we need to as well.