Add a .type getter to VersionNumber class
I have also decided to make VersionType a const object, as these version types are used very often so it would be nice to have an easy way to access them at runtime as well. This can also replace the old Zod-inferred VersionType as the use case for using it as a type is exactly the same.