Make nonNegativeTupleError private static readonly at runtime
It was very cursed because if you log new VersionNumber() to the console, this property still shows up despite the private keyword being used. I think this is because the private keyword is actually an erasable TypeScript compile-time thing, where it only signals to the compiler that it should be private, but it is not enforced at runtime. As such, it can leak out to the terminal output, which... yeah, not great.