Fix the TypeDoc config
I changed it recently such that it needs to be invoked as a function, as ESLint plugin needs it this way so that it can customise the entrypoint. Here we can just invoke it as is with no arguments as the entrypoint defaults to src/index.ts, which is already the entrypoint of this package to begin with.